test: splitted collector/test/tracing/cloud/aws-sdk/v3/sqs into two logical units
